refactor: split MalgoM, RnState and DsState into modules
