+++ malgo.cabal
@@ -40,6 +40,7 @@ library
       Koriel.Scheme.AST
       Malgo.Build
       Malgo.Desugar.DsEnv
+      Malgo.Desugar.DsState
       Malgo.Desugar.Match
       Malgo.Desugar.Pass
       Malgo.Desugar.Type
@@ -54,6 +55,7 @@ library
       Malgo.Lsp.Index
       Malgo.Lsp.Pass
       Malgo.Lsp.Server
+      Malgo.Monad
       Malgo.Parser
       Malgo.Prelude
       Malgo.Refine.Pass
@@ -61,6 +63,7 @@ library
       Malgo.Refine.Space
       Malgo.Rename.Pass
       Malgo.Rename.RnEnv
+      Malgo.Rename.RnState
       Malgo.Syntax
       Malgo.Syntax.Extension
   other-modules:
