+++ src/Malgo/Desugar/DsEnv.hs
@@ -3,22 +3,13 @@
 
 module Malgo.Desugar.DsEnv where
 
-import Control.Lens (mapped, over, traversed, use, (^.), _2)
 import Control.Lens.TH
-import Data.HashMap.Strict qualified as HashMap
-import Data.List qualified as List
-import Koriel.Core.Type qualified as C
 import Koriel.Id
 import Koriel.Lens
 import Koriel.MonadUniq
-import Koriel.Pretty
-import Malgo.Infer.TcEnv (TcEnv)
-import Malgo.Infer.TypeRep
-import Malgo.Infer.TypeRep qualified as GT
-import {-# SOURCE #-} Malgo.Interface (Interface)
+import Malgo.Interface (Interface)
+import Malgo.Monad
 import Malgo.Prelude
-import Malgo.Syntax.Extension
-import qualified Koriel.Core.Syntax as C
 
 -- 脱糖衣処理の環境
 data DsEnv = DsEnv
@@ -32,46 +23,3 @@ makeFieldsNoPrefix ''DsEnv
 
 makeDsEnv :: ModuleName -> MalgoEnv -> DsEnv
 makeDsEnv _moduleName MalgoEnv {..} = DsEnv {..}
-
--- | トップレベル宣言
-data Def
-  = VarDef (Id C.Type) (C.Exp (Id C.Type))
-  | FunDef (Id C.Type) ([Id C.Type], C.Exp (Id C.Type))
-  | ExtDef Text C.Type
-
-makePrisms ''Def
-
-data DsState = DsState
-  { -- | Malgo -> Coreの名前環境
-    _nameEnv :: HashMap RnId (Id C.Type),
-    -- | 型環境
-    _signatureMap :: HashMap RnId (GT.Scheme GT.Type),
-    _typeDefMap :: HashMap RnId (GT.TypeDef GT.Type),
-    _globalDefs :: [Def]
-  }
-
-makeFieldsNoPrefix ''DsState
-
-makeDsState ::
-  TcEnv ->
-  DsState
-makeDsState tcEnv =
-  DsState
-    { _nameEnv = mempty,
-      _signatureMap = tcEnv ^. signatureMap,
-      _typeDefMap = tcEnv ^. typeDefMap,
-      _globalDefs = []
-    }
-
-lookupValueConstructors ::
-  MonadState DsState m =>
-  Id GT.Type ->
-  [GT.Type] ->
-  m [(RnId, Scheme GT.Type)]
-lookupValueConstructors con ts = do
-  typeEnv <- use typeDefMap
-  -- _valueConstructorsがnullのとき、そのフィールドは型シノニムのものなので無視する
-  case List.find (\TypeDef {..} -> _typeConstructor == GT.TyCon con && not (List.null _valueConstructors)) (HashMap.elems typeEnv) of
-    Just TypeDef {..} ->
-      pure $ over (mapped . _2 . traversed) (GT.applySubst $ HashMap.fromList $ zip _typeParameters ts) _valueConstructors
-    Nothing -> errorDoc $ "Not in scope:" <+> quotes (pPrint con)
