+++ src/Malgo/Desugar/Type.hs
@@ -4,7 +4,7 @@ import Koriel.Core.Type
 import Koriel.Core.Type qualified as C
 import Koriel.Id
 import Koriel.Pretty
-import Malgo.Desugar.DsEnv
+import Malgo.Desugar.DsState
 import Malgo.Infer.TypeRep
 import Malgo.Infer.TypeRep qualified as GT
 import Malgo.Prelude
