--- src/Malgo/Interface.hs-boot
@@ -1,3 +1,0 @@
-module Malgo.Interface where
-
-data Interface
