--- src/Malgo/Lsp/Index.hs-boot
@@ -1,3 +1,0 @@
-module Malgo.Lsp.Index where
-
-data Index
